+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 ## v2.0.15
 IMPROVEMENTS:
 * `RelayMode` is now an enum
+* Check TimeZone ranges
 
 ## v2.0.14
 FEATURES: